Attempt to further refine debian bootstrap.
Mostly, without satisfying results:

I don't need systemd and udev installed, yes, I promise.  It installs them anyway.

Several hours in of sporadic fighting (intermingled with enormous amounts of waiting), I can't get the debootstrap process to emit an image smaller than about 280mb.

And the irony of having to apt-get install debootstrap on an ubuntu image that I already lacked good provenance for is just unbearable.

So, we're just going to calmly leave this here, as it stands, and perhaps return to it later.

Signed-off-by: Eric Myhre <hash@exultant.us>