Huzzah!  Finally captured diffoscope.
In a completely non-reproducible way, of course.  Irony is the name of the game, here.

I'm going to hack together a shell script to template the formula of running it.  Subuser would be good at this kind of stuff (it's already got miles of excellent interface design around integrating contained systems back into commands that behave sensibly on the host)... looking forward to that day.

Signed-off-by: Eric Myhre <hash@exultant.us>