Rhone research
If you want to install all of alpine-base, you start getting additional sources of nondeterminism, so here's a strip for that.

We'll maybe not actually follow this train of thought too much further, and just start accumulating individual components we want, without the grabbag and without bringing apk along with into the image -- Using it externally as a factory from pinned sources is quite satisfactory.

Signed-off-by: Eric Myhre <hash@exultant.us>