Formula to capture a recent LTS ubuntu release.
This gathers together all the little patches to make container operation less frictional from all my scattered notes, adds a few things that are (this is, by nature, somewhat opinionated) simply too annoying to live without, and gets that all applied to a source straight from upstream.

And believe it or not?  The results are pretty much deterministic.  That's awfully nice.  (We'll see how that lasts over time.  But between the pre-packaged apt lists and refraining from unreproducible/network-dependent updates to those, we seem to be dodging most of the bullets there...?)

This gives us a nice, new, accountable base image (up to the limit where we punt to upstreams' build farms) -- with the latest critical security vulnerabilities patched.  Hopefully this will remain a safe starting point for quite some time.  (Knock on wood.)

Signed-off-by: Eric Myhre <hash@exultant.us>