Build git.
This is WIP and not fully functioning.  Git builds have... a lot of options.  And it's become clear that running an integration test framework of some kind on the results is quite necessary to determine the validity: this process *does* exit zero/success, but the product doesn't have all the capabilities I'd expect.

Signed-off-by: Eric Myhre <hash@exultant.us>