Swap apk's default scripts for our own.
This was a quest to get permission levels down to policy=uidzero... and we've ALMOST done it!  The apk step would be satisified down there.  However, chroot itself is forbidden until policy>=governor, so we still end up up there.

Though the goal wasn't quite attained, keeping this state anyway, as it's closer.  We *could* split this up into a multi-step build, and use the second step (now also at low policy=uidzero privs) to do the final install steps.  Might be neat.

Signed-off-by: Eric Myhre <hash@exultant.us>