Discard terminfo files.
This is kinda funny: we're actually *so* minimalist here that this is a nontrivial percent change.  2.1M->1.5M -- it's over 25%!

This may bite us later down the line, but I'm inclined to try, and see how long the halflife is before issues crop up.  I think pretty much everyone just defaults to vt200 codes, roughly, this century, don't they?

Signed-off-by: Eric Myhre <hash@exultant.us>