Convert rhone to use reppl, be multi-step, and run entirely at minimal policy=uidzero privs (!! yesss)
Really pleased with this outcome.  Bootstrapping a base image should truly NOT be a priviledged operation in any system.  And for the first time, we have an exhibit in which it properly isn't.

In the process, we shifted our result from 4JkW1xaS0LyN-vNMT2y65LxwPwQC674emhW3P6Q6CuK9d6XpCiZtlnToJkgpXXU7 to L4M-gfAFw1k1z93wMZ2rSNwsbRb-OAGeOZtlCmIKcDWbOGMZ6bNDi41DbhmphTV3.

Why?  Well, let's see:

```
diff \
	<(repeatr examine ware --kind=tar --where=file+ca://./wares/ --hash 4JkW1xaS0LyN-vNMT2y65LxwPwQC674emhW3P6Q6CuK9d6XpCiZtlnToJkgpXXU7) \
	<(repeatr examine ware --kind=tar --where=file+ca://./wares/ --hash L4M-gfAFw1k1z93wMZ2rSNwsbRb-OAGeOZtlCmIKcDWbOGMZ6bNDi41DbhmphTV3)
```

And the answer is:

```
> "./root/"       5       0755    0       0       2010-01-01 00:00:00 +0000 UTC
> "./task/"       5       0755    0       0       2010-01-01 00:00:00 +0000 UTC
```

By introducing the new step, we've caused the Repeatr "cradle" functionality -- the system that makes sure you have a working dir and a user home dir, for minimal sane operation -- to kick in.  We can disable that in our next commit in order to keep things even more pristine (though there's certainly plenty of room to argue that these are reasonable things to exist by default!).

Signed-off-by: Eric Myhre <hash@exultant.us>