Rhone+builder formula with gcc.
We continue to be able to use APK to stack up more packages... entirely from the outside.  I like this.  This is how this kind of tool should work.

Note that we actually end up using the 'rhone' image *twice*: once as the rootfs (it's providing us with bash for stringing things together), and again in another dir, where we mutate it further and then save that result.  This technique is neat: the filesystem segments we're modifying are extremely clear, and the fact that the surrounding "host" environment happens to be a clone of that same filesystem is incidental.

The result filesystem, as gzipped tar (and today, "TEZDCix4JhEACt4GeuqZS24gufNvGbTIc7WifL0nwiKtJ5shogZIZbDq_DXmvcgK"), is an impressively concise 35M.  (A similar quest in ubuntu or debian has resulted in 120~200mb when similarly compressed, so to have gotten something this compact, and with so little effort or grey hairs, is truly nice.)

Signed-off-by: Eric Myhre <hash@exultant.us>