Add libc-dev; builder-test now generates a binary that flies right.
Many lines of commented-out specifics.  This appears to be the minimal items we can bring along from Alpine's ecosystem and still get a result.

Bringing along g++ brings in a *lot* of headers, some of them quite comical (namely, a *ton* of files for... bindings to the java std lib?  Of all things).  Net, it over doubles the gzipped result size, so let's not, without stronger justification.

The libc-dev package in Alpine seems to install the magic 'crti.o' file under the global '/usr/lib/' path.  I'm not one to look a gift horse in the mouth *too* much at this stage, but it's worth noting this might be a caltrop in the future; other distros seem to have started keeping those files in other paths with more metadata, to aid clarity in cross-compiling situations.

Signed-off-by: Eric Myhre <hash@exultant.us>