Add test formulas to examine rhone/alpine support for packages that are commonly considered to be distro-independent and relocatable.  Bad news.
'/lib/ld-musl-x86_64.so.1' might be doing very nearly the same thing as '/lib64/ld-linux-x86-64.so.2' does, but if you don't have exactly that magic string, you're really in trouble.  '/lib64/ld-linux-x86-64.so.2' is baked in to almost every binary I've ever seen, and it's legitimately tough to imagine living effectively without a compatability layer addressing that first step of ELF execution.

Signed-off-by: Eric Myhre <hash@exultant.us>